Start re-establishing customary colors, 1st color
@see https://github.com/ethereum/remix-ide/issues/1481#issuecomment-416374797
This changes the color of a word like bool from the unusual blue
back to the customary pale green.

This is just the first of such changes to illustrate how it would be
done.

As of https://github.com/ethereum/remix-ide/issues/1481#issuecomment-429862126
more suggestions for changing colors back to the customary color-scheme
and adding new colors will be incorporated into subsequent commits.